# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type StudentSignupInputReligion =
  (typeof StudentSignupInputReligion)[keyof typeof StudentSignupInputReligion];

export const StudentSignupInputReligion = {
  Islamic: "Islamic",
  Christian: "Christian",
} as const;

export interface StudentSignupInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 3 */
  studentCode: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
  /**
   * @minimum 1
   * @maximum 12
   */
  grade: number;
  religion: StudentSignupInputReligion;
}

export interface StudentSigninInput {
  studentCode: string;
  password: string;
}

export interface TeacherSigninInput {
  fullName: string;
  subject: string;
  subjectPassword: string;
}

export type UserInfoRole = (typeof UserInfoRole)[keyof typeof UserInfoRole];

export const UserInfoRole = {
  student: "student",
  teacher: "teacher",
} as const;

export interface UserInfo {
  /** @nullable */
  id?: number | null;
  fullName?: string;
  /** @nullable */
  studentCode?: string | null;
  /** @nullable */
  grade?: number | null;
  /** @nullable */
  religion?: string | null;
  /** @nullable */
  subject?: string | null;
  role: UserInfoRole;
}

export interface AuthResponse {
  token: string;
  user: UserInfo;
}

export type MeResponseRole =
  (typeof MeResponseRole)[keyof typeof MeResponseRole];

export const MeResponseRole = {
  student: "student",
  teacher: "teacher",
} as const;

export interface MeResponse {
  /** @nullable */
  id?: number | null;
  fullName?: string;
  /** @nullable */
  studentCode?: string | null;
  /** @nullable */
  grade?: number | null;
  /** @nullable */
  religion?: string | null;
  /** @nullable */
  subject?: string | null;
  role: MeResponseRole;
}

export interface TeacherFile {
  id: number;
  filename: string;
  filepath: string;
  subject: string;
  uploadedBy: string;
  createdAt: string;
}

export interface FileUploadInput {
  subject: string;
}

export interface Student {
  id: number;
  fullName: string;
  studentCode: string;
  grade: number;
  religion: string;
  createdAt: string;
}

export type RatingLabel = (typeof RatingLabel)[keyof typeof RatingLabel];

export const RatingLabel = {
  Excellent: "Excellent",
  Good: "Good",
  Average: "Average",
  Needs_Improvement: "Needs Improvement",
} as const;

export interface Rating {
  id: number;
  studentId: number;
  subject: string;
  teacherName: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  label: RatingLabel;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  studentName?: string | null;
  createdAt: string;
}

export interface RatingInput {
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  comment: string;
}

export interface Notebook {
  id: number;
  studentId: number;
  subject: string;
  notebookName: string;
  createdAt: string;
}

export interface NotebookInput {
  subject: string;
  notebookName: string;
}

export interface NotebookUpdate {
  notebookName: string;
}

export interface Note {
  id: number;
  notebookId: number;
  content: string;
  createdAt: string;
}

export interface NoteInput {
  notebookId: number;
  content: string;
}

export interface NoteUpdate {
  content: string;
}

export interface NoteFile {
  id: number;
  notebookId: number;
  filename: string;
  filepath: string;
  createdAt: string;
}

export interface NoteFileUploadInput {
  notebookId: number;
}

export interface SummarizeInput {
  fileId: number;
}

export interface SummaryResult {
  mainTopic: string;
  keyPoints: string[];
  importantNotes: string[];
}

export interface FlashcardsInput {
  notebookIds: number[];
}

export interface Flashcard {
  question: string;
  answer: string;
}

export interface FlashcardsResult {
  flashcards: Flashcard[];
}

export type GetFilesParams = {
  subject: string;
};

export type GetStudentsParams = {
  search?: string;
};

export type GetNotebooksParams = {
  subject: string;
};

export type GetNotesParams = {
  notebookId: number;
};

export type GetNoteFilesParams = {
  notebookId: number;
};