# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learnova LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/student/signup:
    post:
      operationId: studentSignup
      tags: [auth]
      summary: Student registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentSignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Student code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/student/signin:
    post:
      operationId: studentSignin
      tags: [auth]
      summary: Student login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentSigninInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/teacher/signin:
    post:
      operationId: teacherSignin
      tags: [auth]
      summary: Teacher login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherSigninInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user from token
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Files ─────────────────────────────────────────────────────────────────
  /files:
    get:
      operationId: getFiles
      tags: [files]
      summary: List files for a subject (teacher or student)
      parameters:
        - in: query
          name: subject
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherFile"
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file (teacher only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherFile"
        "400":
          description: Invalid file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file (teacher only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Students ──────────────────────────────────────────────────────────────
...
  /ai/summarize:
    post:
      operationId: summarizeFile
      tags: [ai]
      summary: Summarize a teacher file using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummarizeInput"
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResult"
        "400":
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/flashcards:
    post:
      operationId: generateFlashcards
      tags: [ai]
      summary: Generate flashcards from selected notebooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardsInput"
      responses:
        "200":
          description: Generated flashcards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardsResult"
        "400":
          description: No content found in notebooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth schemas ──────────────────────────────────────────────────────
    StudentSignupInput:
      type: object
      required: [fullName, studentCode, password, confirmPassword, grade, religion]
      properties:
        fullName:
          type: string
          minLength: 2
        studentCode:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        grade:
          type: integer
          minimum: 1
          maximum: 12
        religion:
          type: string
          enum: [Islamic, Christian]
    StudentSigninInput:
      type: object
      required: [studentCode, password]
      properties:
        studentCode:
          type: string
        password:
          type: string
    TeacherSigninInput:
      type: object
      required: [fullName, subject, subjectPassword]
      properties:
        fullName:
          type: string
        subject:
          type: string
        subjectPassword:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      required: [role]
      properties:
        id:
          type: ["integer", "null"]
        fullName:
          type: string
        studentCode:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        religion:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, teacher]
    MeResponse:
      type: object
      required: [role]
      properties:
        id:
          type: ["integer", "null"]
        fullName:
          type: string
        studentCode:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        religion:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, teacher]

    # ── File schemas ──────────────────────────────────────────────────────
    TeacherFile:
      type: object
      required: [id, filename, filepath, subject, uploadedBy, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        filepath:
          type: string
        subject:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    FileUploadInput:
      type: object
      required: [subject]
      properties:
        subject:
          type: string

    # ── Student schemas ───────────────────────────────────────────────────
...
    SummarizeInput:
      type: object
      required: [fileId]
      properties:
        fileId:
          type: integer
    SummaryResult:
      type: object
      required: [mainTopic, keyPoints, importantNotes]
      properties:
        mainTopic:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        importantNotes:
          type: array
          items:
            type: string
    FlashcardsInput:
      type: object
      required: [notebookIds]
      properties:
        notebookIds:
          type: array
          items:
            type: integer
    Flashcard:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
    FlashcardsResult:
      type: object
      required: [flashcards]
      properties:
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"