# lib/api-zod/src/generated/types/flashcardsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from "./flashcard";

export interface FlashcardsResult {
  flashcards: Flashcard[];
}