# lib/api-zod/src/generated/types/studentSignupInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentSignupInputReligion } from "./studentSignupInputReligion";

export interface StudentSignupInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 3 */
  studentCode: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
  /**
   * @minimum 1
   * @maximum 12
   */
  grade: number;
  religion: StudentSignupInputReligion;
}