# lib/api-zod/src/generated/types/summaryResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SummaryResult {
  mainTopic: string;
  keyPoints: string[];
  importantNotes: string[];
}